Moving reused components to core
