+++ views/core/components/ErrorBoundary.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import Error from '~/views/home/components/Error'
+import Error from '~/views/core/components/Error'
 
 class ErrorBoundary extends React.Component {
   constructor(props) {
