+++ views/core/components/SlideIterator.js
